Mean and Median of airquality datasets variables.
4. Mr. Cucumber has a shop selling fruit and vegetables called Green and Good. He is thinking
about how the new kind of pumpkin sells in the shop. During 30 days, he observes the sales
pattern of the new pumpkin. These are his observations:
pumpkins=data.frame(no_of_pumpkins=c(0,1,2,3,4,5,6,7),
days=c(10,2,3,5,4,2,4,5))
Determine and calculate the arithmetic mean of the number of sold pumpkins.
5. The package datasets contains a lot of different data sets. The data set presents daily air
quality measurements in New York, May to September 1973.
(a) Load the package datasets.
(b) Calculate the arithmetic mean of the temperature.
(c) Calcullate the median of the solar radiation.
(d) In the previous task, you encountered a problem in the calculation of the median. Which
problem? Recalculate the median of the solar radiation, taking this problem into account.