Correlation, Scatter plot, Outlier of data
1. These data show the prices for food in two different supermarkets. The first item in the vector
for supermarket A corresponds to the first item in the vector for supermarket B and so on.
A=c(1, 5, 6.6, 4, 10, 12, 13, 21, 1, 3)
and
B=c(1.2, 57, 4.6, 4.1, 10.4, 12.9, 11.9, 22, 1.4, 4)
(a) Calculate the correlation of the prices in the two supermarkets. Interpret the correlation.
(b) Construct a scatter plot for the data.
(c) Does the scatter plot confirm the result in a)?
(d) Remove the outlier and perform the tasks a) - c) again!