Histogram and Skewness.
(a) How many cities participated in the study?
(b) Create a histogram of the number of citizens. What can you say
about the skewness of the data? (symmetric, left-skewed, rightskewed)
(c) Create a histogram with about 20 bins of the number of citizens
(use breaks= in the hist-command). What can you now say about
the skewness of the data? (symmetric, left-skewed, right-skewed)