Barplot, Histogram, Piechart and frequencies
a) Use the function table to summarize the data. Which it the most
common number of rooms? How many flats have this number of
rooms?
(b) Use the function barplot to illustrate the data.
(c) Calculate the relative frequencies (= frequencies expressed as percent) of the number of rooms.
(d) Use a pie chart to illustrate the relative frequencies of the number
of rooms. Function: pie