Barplot, boxplot and data analysis
(a) Create the data set hospital data as described above.
(b) What are the dimensions of the data set hospital data?
(c) Explain the scale of each of the variables!
(d) Create for each of the following variables an appropriate diagram
and motivate the choice of diagram:
i. age
ii. gender
iii. degree
iv. stay
v. diagnosis
(e) Create and compare the boxplots for the variable stay for females
and males!
(f) Use the function quantile to calculate the three quantiles used in
5e! Compare these quantiles for females and males!