Hypothesis test using t.test.
1. In the data set beaver1 in the package datasets a small part of a study
of the long-term temperature dynamics of beaver Castor canadensis in
north-central Wisconsin is contained. Body temperature was measured
by telemetry every 10 minutes for a female beaver.
(a) Load the data set beaver1 in the package datasets by giving the
command library(datasets).
(b) Use the function t.test to perform a two-sided test for the population mean of the variable temp with null hypothesis mean=38.0.
Interpret the results. Can you reject the null hypothesis? What
is the p-value? Interpret the p-value.
(c) Use the function t.test to perform a two-sided test for the population mean of the variable temp with null hypothesis mean=37.0.
Interpret the results. Can you reject the null hypothesis? What
is the p-value? Interpret the p-value.
(d) Use the function t.test to perform a two-sided test for the population mean of the variable temp with null hypothesis mean=36.9.
Interpret the results. Can you reject the null hypothesis? What
is the p-value? Interpret the p-value.
(e) Use the function t.test to perform a two-sided test for the population mean of the variable temp with null hypothesis mean=36.86.
Interpret the results. Can you reject the null hypothesis? What
is the p-value? Interpret the p-value.
(f) Compare the results and explain the differences in 1b, 1c, 1d and
1e!